refactor: entities/meeting store/types 통합 및 import 경로 정리
