+++ src/entities/meetings/model/types.ts
