+++ src/views/home/HomePage.tsx
@@ -3,8 +3,8 @@
 import { useMemo } from 'react';
 import { DrinkType } from '@/shared';
 import { useUserStore } from '@/entities/user';
-import { HomeListKind } from '@/entities/meetings/types';
 import { HomeMeetingSection, HomeLoginOverlay } from '@/widgets/home';
+import { HomeListKind } from '@/entities/meetings';
 
 export default function HomePage() {
   const rawInterests = useUserStore((s) => s.user?.interests);
