+++ src/widgets/home/HomeMeetingSection.tsx
@@ -2,9 +2,8 @@
 
 import { DrinkType } from '@/shared';
 import { useUserStore } from '@/entities/user';
-import { HomeListKind, MeetingCardInfo } from '@/entities/meetings/types';
 import { MeetingCardCarousel } from '@/widgets';
-import { useHomeMeetingsQuery } from '@/entities/meetings';
+import { HomeListKind, MeetingCardInfo, useHomeMeetingsQuery } from '@/entities/meetings';
 
 function emptyCopy(kind: HomeListKind) {
   switch (kind) {
